Expo: Cast denominators of integer divisions to floats before performing a       matrix translation to viewports with the results. Otherwise rounding       errors in cases the results are not an integers cause the viewports to       be painted offscreen.
Fixes LP: #438580

(bzr r3753.4.1)

+++ plugins/expo/src/expo.cpp
@@ -954,8 +954,8 @@ ExpoScreen::paintWall (const GLScreenPaintAttrib &attrib,
 
 	    sTransform3 = sTransform2;
 
-	    sTransform3.translate ( output->x () / output->width (),
-				   -output->y () / output->height (), 0.0);
+	    sTransform3.translate ( output->x () / static_cast <float> (output->width ()),
+				   -output->y () / static_cast <float> (output->height ()), 0.0);
 
 	    cScreen->setWindowPaintOffset ((screen->vp ().x () - i) *
 					   screen->width (),
